feat: add mysql chart (#5)
I ran the following to pull the bitnami mysql chart and index it. I plan
to use this instead of the mariadb chart as we are upgrading to mysql 8.

```
helm pull oci://registry-1.docker.io/bitnamicharts/mysql
./scripts/update-index.sh
```

The mariadb version is indexed with a different version suffixed by
`-subpath`. I reverted those changes to the index file to avoid breaking
changes as they appear to be custom.

QSS-2315